refactor(adoc-core): unit 8 — artifact/ → infrastructure/artifact/ + pivot AgentJsonDocument to domain
Remove data type defs from infrastructure/artifact/agent_json.rs; they
now live at domain/artifact.rs. Remove ArtifactWriter trait def from
artifact/mod.rs (deduped to domain/ports/artifact_writer.rs). Pivot
lib.rs re-export from artifact::agent_json to domain::artifact.

Refs #27.
