+++ crates/adoc-core/src/infrastructure/mod.rs
@@ -1,3 +1,4 @@
 // see ADR-0009
+pub(crate) mod artifact;
 pub(crate) mod render;
 pub(crate) mod source;
