+++ crates/adoc-core/src/lib.rs
@@ -1,11 +1,10 @@
 mod application;
-mod artifact;
 mod compile;
 mod domain;
 mod infrastructure;
 mod parser;
 mod validate;
 
-pub use artifact::agent_json::AgentJsonDocument;
+pub use domain::artifact::AgentJsonDocument;
 pub use compile::{BuildArtifacts, CompileInput, CompileResult, compile_workspace};
 pub use domain::diagnostic::{Diagnostic, DiagnosticCode, Severity};
